agent/requirements.txt
# Slim Quality Agent - Dependências Python
# Última atualização: 2024-12-28
# Stack: LangGraph + FastAPI + Supabase + SICC
# Versões: Dezembro 2024 (testadas e validadas)

# Framework Web
fastapi==0.115.0
uvicorn[standard]==0.32.0

# LangChain/LangGraph (versões atuais Dez 2024)
# LangGraph 1.0.5 - Versão estável atual (Dez 2024)
# LangChain 1.2.0 - Versão atual compatível
langchain==1.2.0
langgraph==1.0.5
langchain-anthropic==1.3.0
langchain-core==1.2.5

# OpenAI API
openai>=1.50.0

# Supabase Client
supabase==2.27.0

# Cache e Storage
redis>=5.0.0
httpx  # Deixar pip resolver versão compatível

# Utilitários
python-dotenv>=1.0.0
pydantic>=2.11.0
pydantic-settings>=2.5.0
python-multipart>=0.0.9

# MCP SDK
mcp>=1.25.0

# Google APIs (Biblioteca atual disponível)
google-api-python-client>=2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-generativeai>=0.8.0  # Biblioteca atual (deprecated mas funcional)

# Logging
structlog>=23.2.0

# ============================================
# SICC - Sistema de Inteligência Corporativa
# ============================================

# Embeddings e Processamento de Texto
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Banco de Dados e Vetores
psycopg2-binary>=2.9.7
pgvector>=0.2.4
sqlalchemy>=2.0.0
alembic>=1.12.0

# Processamento Assíncrono
asyncio-mqtt>=0.13.0
aioredis>=2.0.1
celery>=5.3.0

# Análise de Dados e Métricas
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testes Específicos SICC
hypothesis>=6.88.0  # Property-based testing

# Desenvolvimento
pytest>=7.4.0
pytest-asyncio>=0.21.0